applied PR suggestion
