+++ app/uk/gov/hmrc/cataloguefrontend/metrics/model/Group.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.cataloguefrontend.connector.model
+package uk.gov.hmrc.cataloguefrontend.metrics.model
 
 //todo: should it be called DependencyGroup or similar?
 case class Group(name: GroupName, repos: Seq[RepositoryName], dependencies: Seq[DependencyName])
