+++ app/uk/gov/hmrc/cataloguefrontend/metrics/model/GroupName.scala
@@ -14,13 +14,6 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.cataloguefrontend.connector.model
+package uk.gov.hmrc.cataloguefrontend.metrics.model
 
-case class DependencyName (value: String) extends AnyVal
-
-object DependencyName {
-  def apply(serviceProgressMetricses: Seq[ServiceProgressMetrics]): Seq[DependencyName] = serviceProgressMetricses
-    .map(_.name)
-    .distinct
-    .map(DependencyName.apply)
-}
+final case class GroupName(value: String)
