+++ app/uk/gov/hmrc/cataloguefrontend/metrics/views/MetricsExplorerPage.scala.html
@@ -15,9 +15,8 @@
 *@
 
 @import play.api.mvc.Call
-@import uk.gov.hmrc.cataloguefrontend.connector.model.{Group, Repository, DependencyName}
+@import uk.gov.hmrc.cataloguefrontend.metrics.model.{Group, Repository, DependencyName, MetricsEntry}
 @import uk.gov.hmrc.cataloguefrontend.MetricsExplorerController
-@import uk.gov.hmrc.cataloguefrontend.connector.model.MetricsEntry
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages
 @import uk.gov.hmrc.cataloguefrontend.{ routes => appRoutes }
 @import uk.gov.hmrc.cataloguefrontend.service.SearchByUrlService.FrontendRoutes
